+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Additions and Fixes:
 - Added error handling to all slash domains (previously only t3, mod and manager had it) ([#139](https://github.com/crscillitoe/RoboBanana/pull/139))
 - Added auditing of streamdeck prediction endpoints ([#140](https://github.com/crscillitoe/RoboBanana/pull/140))
+- Added marker commands to easily allow multiple people to create stream markers for editors and VOD channel ([#141](https://github.com/crscillitoe/RoboBanana/pull/141))
   
 ## 2024.04.04
 This update requires `/sync sync` to be run once